fix(backend): corregi hallazgos bloqueantes de la retroalimentacion (DATABASE_URL, lifespan, Numeric, seed idempotente)

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, Numeric
 from app.database import Base
 
 class Product(Base):
@@ -6,7 +6,7 @@ class Product(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
     description = Column(String, nullable=False)
     category = Column(String, nullable=False, index=True)
     image = Column(String, nullable=False)
